refactor: type defs
